// apps/events-admin/src/features/calendar/interfaces.ts
import type { TEventColor } from "@/features/calendar/types";

export interface IUser {
	id: string;
	name: string;
	picturePath: string | null;
}

export interface IEvent {
	id: number;
	startDate: string;
	endDate: string;
	title: string;
	color: TEventColor;
	description: string;
	user: IUser;
    isAllDay?: boolean;
}

export interface ICalendarCell {
	day: number;
	currentMonth: boolean;
	date: Date;
}